HexagonUtils updated to work with all positions. It uses modulu calculations to ensure no ArrayIndexOutOfBounds
Grass is implemented and grows based of waterAmmount in neighbouring
tiles. To much water and the grass dies, to litte water and the grass
dies, with just the right aount it grows.

Water is implemented however it returns a static 0.5f on groundLevel,
Clouds and rain is yet to be implemented